+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Unreleased 0.12.0
 
 * Pull updates in cryptostore, fixing build with crypton.
+* Stabilize ordering of keys, both in the header and in claims, resolving
+  [#2](https://github.com/puffnfresh/haskell-jwt/issues/2).
+* Fix and reenable doctests.
 
 # 2021-12-11 0.11.0